feat: Add ThemeSelect and TranslationSelect components
This commit adds the ThemeSelect and TranslationSelect components to the project. These components are used for selecting the theme and translation language respectively. They provide a dropdown interface with options for light/dark theme and English/Indonesia translation. These components will be used for enhancing the user experience and customization options in the application.