feat: Update name field in auth POST route to dynamically set user role
The name field in the auth POST route has been updated to dynamically set the user role based on the email address. If the email address is 'admin@gmail.com', the name field is set to 'Admin' and the role field is set to 'admin'. For any other email address, the name field is set to 'Guest' and the role field is set to 'guest'. This change improves the accuracy of user role assignment and enhances the functionality of the authentication process.

Co-authored-by: [Author Name]